+++ connector-out.php
@@ -15,7 +15,10 @@ require_once __DIR__ . '/vendor/autoload.php';
 use Camunda\Entity\Request\ExternalTaskRequest;
 use Camunda\Service\ExternalTaskService;
 use PhpAmqpLib\Connection\AMQPStreamConnection;
+use PhpAmqpLib\Exception\AMQPRuntimeException;
 use Quancy\Logger\Logger;
+
+const WAIT_BEFORE_RECONNECT_US = 1000;
 
 // Config
 $config = __DIR__ . '/config.php';
@@ -63,12 +66,69 @@ function bad_work($externalTaskService, $message) {
 };
 
 /**
+ * Validate message
+ */
+function validate_message($message) {
+    // Headers
+    if(!isset($message['headers'])) {
+        $message = '`headers` not is set in incoming message';
+        Logger::log($message, 'output', RMQ_QUEUE_OUT,'bpm-connector-out', 1);
+        exit(1);
+    }
+
+    // Unsafe parameters in headers
+    $unsafeHeadersParams = ['camundaWorkerId', 'camundaExternalTaskId'];
+
+    foreach ($unsafeHeadersParams as $paramName) {
+        if(!isset($message['headers'][$paramName])) {
+            $message = '`' . $paramName . '` param not is set in incoming message';
+            Logger::log($message, 'output', RMQ_QUEUE_OUT,'bpm-connector-out', 1);
+            exit(1);
+        }
+    }
+}
+/**
+ * Close connection
+ *
+ * @param $connection
+ */
+function cleanup_connection($connection) {
+    // Connection might already be closed.
+    // Ignoring exceptions.
+    try {
+        if($connection !== null) {
+            $connection->close();
+        }
+    } catch (\ErrorException $e) {
+    }
+}
+
+/**
+ * Shutdown
+ *
+ * @param $connection
+ */
+function shutdown($connection)
+{
+    //$channel->close();
+    $connection->close();
+}
+
+
+/**
  * Callback
  *
  * @param $msg
  */
 $callback = function($msg) {
     Logger::log(sprintf("Received %s", $msg->body), 'input', RMQ_QUEUE_OUT,'bpm-connector-out', 0 );
+
+    // Set manual acknowledge for received message
+    $msg->delivery_info['channel']->basic_ack($msg->delivery_info['delivery_tag']); // manual confirm delivery message
+    // Send a message with the string "quit" to cancel the consumer.
+    if ($msg->body === 'quit') {
+        $msg->delivery_info['channel']->basic_cancel($msg->delivery_info['consumer_tag']);
+    }
 
     // Update variables
     $message = json_decode($msg->body, true);
@@ -80,7 +140,11 @@ $callback = function($msg) {
     ];
 
     // Request to Camunda
-    $externalTaskService = new ExternalTaskService(CAMUNDA_API_URL);
+    $camundaUrl = sprintf(CAMUNDA_API_URL, CAMUNDA_API_LOGIN, CAMUNDA_API_PASS); // camunda api with basic auth
+    $externalTaskService = new ExternalTaskService($camundaUrl);
+
+    // Validate message
+    validate_message($message);
 
     // Сomplete task if his status is success
     // and retry it if is not succcess
@@ -96,26 +160,39 @@ $callback = function($msg) {
 
 };
 
-// Open connection
-$connection = new AMQPStreamConnection(RMQ_HOST, RMQ_PORT, RMQ_USER, RMQ_PASS, RMQ_VHOST, false, 'AMQPLAIN', null, 'en_US', 3.0, 3.0, null, true, 60);
-$channel = $connection->channel();
-$channel->confirm_select(); // change channel mode to confirm mode
-$channel->queue_declare(RMQ_QUEUE_OUT, false, true, false, false);
-$channel->basic_consume(RMQ_QUEUE_OUT, '', false, true, false, false, $callback);
+/**
+ * Loop
+ */
+$connection = null;
+while(true) {
+    try {
+        $connection = new AMQPStreamConnection(RMQ_HOST, RMQ_PORT, RMQ_USER, RMQ_PASS, RMQ_VHOST, false, 'AMQPLAIN', null, 'en_US', 3.0, 3.0, null, true, 60);
+        register_shutdown_function('shutdown', $connection);
 
-// Variate timeout for reboot worker in random time
-$timeout = mt_rand(36000, 50400);
+        echo ' [*] Waiting for messages. To exit press CTRL+C', "\n";
 
-Logger::log('Waiting for messages. To exit press CTRL+C', 'output', RMQ_QUEUE_OUT,'bpm-connector-out', 0);
+        // Your application code goes here.
+        $channel = $connection->channel();
+        $channel->confirm_select(); // change channel mode to confirm mode
+        $channel->basic_qos(0, 1, false); // one message in one loop
+        $channel->basic_consume(RMQ_QUEUE_OUT, '', false, false, false, false, $callback);
 
-try {
-    while (count($channel->callbacks)) {
-        $channel->wait(null, false, $timeout);
+        while ($channel->is_consuming()) {
+            $channel->wait(null, true, 0);
+            sleep(1);
+        }
+
+    } catch(AMQPRuntimeException $e) {
+        echo $e->getMessage() . PHP_EOL;
+        cleanup_connection($connection);
+        usleep(WAIT_BEFORE_RECONNECT_US);
+    } catch(\RuntimeException $e) {
+        echo "Runtime exception " . PHP_EOL;
+        cleanup_connection($connection);
+        usleep(WAIT_BEFORE_RECONNECT_US);
+    } catch(\ErrorException $e) {
+        echo "Error exception " . PHP_EOL;
+        cleanup_connection($connection);
+        usleep(WAIT_BEFORE_RECONNECT_US);
     }
-} catch(Exception $e) {
-    Logger::log('Planned reboot by timeout.', 'output', RMQ_QUEUE_OUT,'bpm-connector-out', 0 );
 }
-
-// Close connection
-$channel->close();
-$connection->close();